requirements.txt
# Core dependencies
aiohttp==3.9.1
asyncio==3.4.3
redis==5.0.1
pyyaml==6.0.1

# Network libraries
aiohttp-cors==0.7.0
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
locust==2.19.1

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
colorlog==6.8.0
